pyproject.toml
[project]
name = "pyutils"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "lijc210", email = "lijc210@163.com" }]
dependencies = [
    "apscheduler~=3.10.1",
    "clickhouse-driver==0.2.6",
    "fastapi==0.70.1",
    "ldap3==2.9.1",
    "numpy<2",
    # "psycopg2-binary==2.9.5",
    "PyMySQL==1.0.2",
    "pyodps==0.11.6.1",
    "pyquery==2.0.0",
    "python-dateutil==2.8.2",
    "python-multipart==0.0.5",
    "redis==3.5.3",
    "requests==2.26.0",
    "retry==0.9.2",
    "sentry-sdk~=1.21.0",
    "uvicorn==0.16.0",
    "huaweicloudsdkdli==3.1.19",
    "redis-py-cluster==2.1.3",
    "DBUtils==3.0.2",
    "pyquery~=2.0.0",
    "imbox~=0.9.8",
    "happybase~=1.2.0",
    "easybase~=1.1.3",
    "urllib3~=1.26.15",
    "chardet~=4.0.0",
    "func_timeout~=4.3.5",
    # "line_profiler~=4.0.3",
    "fire>=0.5.0",
    "apache-airflow==2.7.0",
    "airflow-code-editor==7.4.1",
    "esdk-obs-python>=3.23.9.1",
    "alibabacloud_ecs20140526>=3.0.12",
    "alibabacloud_dataworks_public20200518>=4.7.2",
    "pykafka>=2.8.0",
    "playwright~=1.41.2",
    # pymssql 在macOS M 芯片的依赖
    # brew install freetds
    # brew install openssl
    # "pymssql==2.2.7",
    "alibabacloud-maxcompute20220104>=2.3.0",
    "pandas>=2.2.2",
    "openpyxl>=3.1.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pyutils"]
